80X load time improvement
Moved hashtable generating code to its own executable program, which
makes the hashtable much like before except in contiguous memory and
then dumps this memory to a file. Code in hashtable.asm reads from this
memory dump and corrects the pointers in the hash table if necessary.
dictionary.txt was changed to include null bytes after every word, so
it could be sucked up into memory all at once.
More and better comments forthcoming in the next few days or so.