Shuffle, restart, and access to time funct
Added the shuffle function (you guys may want to look over it because I very well could have made a mistake).  The restart function has been added.  I also added the time function to the table and access to it.  If we want to have a score display function or something of that nature it needs to be added to the help command as well as to the function table.

Email me if you guys think that i messed up on the code.  Thanks